Updated physics of entities. Torches now burn out even in the inventory. Additions to combat.

+++ src/main/java/tgw/evolution/items/ItemClock.java
@@ -19,7 +19,7 @@ public class ItemClock extends ItemEv {
     public ActionResult<ItemStack> onItemRightClick(World worldIn, PlayerEntity playerIn, Hand handIn) {
         ItemStack stack = playerIn.getHeldItem(handIn);
         if (!worldIn.isRemote) {
-            FullDate date = new FullDate((int) worldIn.getDayTime());
+            FullDate date = new FullDate(worldIn.getDayTime());
             playerIn.sendMessage(new StringTextComponent(date.getFullString()));
         }
         return new ActionResult<>(ActionResultType.SUCCESS, stack);
